Update: 将属于图表的 options 转移至 charts_options (#1099)
